+++ vite.config.ts
@@ -1,6 +1,13 @@
 // vite.config.ts
 import { defineConfig } from 'vite';
 
+import { createHtmlPlugin } from 'vite-plugin-html';
+
 export default defineConfig({
   //
+  plugins: [
+    createHtmlPlugin({
+      minify: false,
+    }),
+  ],
 });