+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Private directory
 - Added PostCSS and Plugins
 - Implement Custom Media properties
+- Added Vite Plugin for HTML minification
 
 ## v1.4.3 (Sep 6, 2022)